Create Bat model, add psycopg2 (#16)
* Add Bat model

Include some fields that will be used to keep track of information about
bats. This includes migrations, and an admin interface to edit and add bats.

* Add psycopg2 dependency

Since we'll be using PostgreSQL as our main database driver, we need to
have the driver installed for it.

* Add ModelAdmin to INSTALLED_APPS

This lets us define custom admin pages based off of our models.