Fix API routing (#25)
In order to prevent drf_yasg auto generating routes for Wagtail URLs, a
urlconf must be specified when calling get_schema_view. If we generate
the URL earlier while routing, then all the base URLs generated (for
Redoc) will be relative to that URL.

To mitigate this, the urlpatterns for the version will generate the API
for that version within its own urls.py file, instead of being generated
in the main application. This allows get_schema_view to see that
urlpattern and base all of the URLs from there, instead of from blank.

This is not intuitive, hence why additional comments were left to
explain why. Also, we kept the docs underneath the specifically versioned
one, but don't show up underneath calling the API directly, as to not
have that route (docs/) in the documentation.