Create user registration over API (#26)
This change exposes authentication URLs underneath `/auth/` for the
application, borrowing on Djoser as a template. A custom view was
created in order to show the paths for authentication.

tox.ini was modified in order to force the environment to be rebuilt
if any of the requirements changed, in both base.txt or development.txt.
It so happened that since development did not change while base did, 
the environment had to be force remade to query the new dependencies.